// HelpDeskFE/HelpDeskFETp/src/api/model/createSupplierDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateSupplierDto {
  idOrganization: number;
  /**
     * @minLength 1
     * @maxLength 500
     */
  description: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  contact: string;
  /**
     * @maxLength 240
     * @nullable
     */
  email?: string | null;
  /**
     * @maxLength 13
     * @nullable
     */
  phone?: string | null;
  registrationDate: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  address: string;
  compliance: boolean;
  isActive: boolean;
}